+++ .travis.yml
@@ -13,8 +13,8 @@ install:
   - pip install -U pytest pytest-cov codecov pycodestyle
 
 script:
-  - pytest tests/ --cov=./multiview/
-  - pycodestyle multiview/
+  - pytest tests/ --cov=./mvlearn/
+  - pycodestyle mvlearn/
 
 after_success:
   - codecov
