+++ README.md
@@ -1,4 +1,4 @@
-# multiview
+# mvlearn
 
 [![Python](https://img.shields.io/badge/python-3.7-blue.svg)]()
 [![Build Status](https://travis-ci.com/NeuroDataDesign/multiview.svg?branch=master)](https://travis-ci.com/NeuroDataDesign/multiview)
@@ -8,4 +8,4 @@
 
 
 
-multiview is a Python module for multiview learning.
+mvlearn is a Python module for multiview learning.
