+++ mvlearn/cotraining/base.py
