+++ mvlearn/embed/base.py
