+++ mvlearn/embed/kcca.py
