+++ mvlearn/embed/omnibus.py
