+++ mvlearn/embed/splitae.py
